pi-swarm/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
llama-cpp-python>=0.3.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
httpx>=0.27.0
pyyaml>=6.0
jinja2>=3.1.0
aiosqlite>=0.20.0
huggingface-hub>=0.25.0
